+++ source/AirPlane.h
@@ -30,9 +30,9 @@ public:
 	const int OUT_END_POS_Y = IN_START_POS_Y;
 	const int ROLL_POINT_ADJUST_X = 200;
 
-	const Vector3 IN_START_POS = Vector3(IN_START_POS_X, IN_START_POS_Y, 0.0f);
-	const Vector3 STAY_POS = Vector3(STAY_POS_X, STAY_POS_Y, 0.0f);
-	const Vector3 OUT_END_POS = Vector3(OUT_END_POS_X, OUT_END_POS_Y, 0.0f);
+	const Vector3 IN_START_POS = Vector3( static_cast<float>( IN_START_POS_X ), static_cast<float>( IN_START_POS_Y ), 0.0f );
+	const Vector3 STAY_POS = Vector3( static_cast<float>( STAY_POS_X ), static_cast<float>( STAY_POS_Y ), 0.0f );
+	const Vector3 OUT_END_POS = Vector3( static_cast<float>( OUT_END_POS_X ), static_cast<float>( OUT_END_POS_Y ), 0.0f );
 
 	enum
 	{
